Convert AgendaReducer
